mcp/crash-course/requirements.txt
mcp[cli]
openai
python-dotenv
ipykernel
httpx
